+++ pkg/statsd/handler_cloud_test.go
@@ -301,51 +301,6 @@ func se2() gostatsd.Event {
 	}
 }
 
-type nopHandler struct{}
-
-func (nh *nopHandler) EstimatedTags() int {
-	return 0
-}
-
-func (nh *nopHandler) DispatchMetric(ctx context.Context, m *gostatsd.Metric) error {
-	return nil
-}
-
-func (nh *nopHandler) DispatchEvent(ctx context.Context, e *gostatsd.Event) error {
-	return nil
-}
-
-func (nh *nopHandler) WaitForEvents() {
-}
-
-type countingHandler struct {
-	mu      sync.Mutex
-	metrics []gostatsd.Metric
-	events  gostatsd.Events
-}
-
-func (ch *countingHandler) EstimatedTags() int {
-	return 0
-}
-
-func (ch *countingHandler) DispatchMetric(ctx context.Context, m *gostatsd.Metric) error {
-	m.DoneFunc = nil // Clear DoneFunc because it contains non-predictable variable data which interferes with the tests
-	ch.mu.Lock()
-	defer ch.mu.Unlock()
-	ch.metrics = append(ch.metrics, *m)
-	return nil
-}
-
-func (ch *countingHandler) DispatchEvent(ctx context.Context, e *gostatsd.Event) error {
-	ch.mu.Lock()
-	defer ch.mu.Unlock()
-	ch.events = append(ch.events, *e)
-	return nil
-}
-
-func (ch *countingHandler) WaitForEvents() {
-}
-
 type CountingProvider interface {
 	Invocations() uint64
 }